+++ CHANGES.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Litmusify: `-litmus` now executes litmus on the remote machine.
 - Added new `act tool` command, which will eventually allow free-form running
   of the various tools act wraps.  Currently it supports herd
   (`act tool herd`), but doesn't do anything special.